Created events for Request.vue
