+++ sources/packages/web/src/views/aest/AssessmentsSummary.vue
@@ -7,12 +7,16 @@
     }"
     subTitle="Assessment"
   />
-  <RequestAssessment :applicationId="applicationId" :studentId="studentId" />
+  <RequestAssessment
+    :applicationId="applicationId"
+    @viewStudentAppeal="goToStudentAppeal"
+  />
   <HistoryAssessment :applicationId="applicationId" />
 </template>
 <script lang="ts">
 import HeaderNavigator from "@/components/generic/HeaderNavigator.vue";
 import { AESTRoutesConst } from "@/constants/routes/RouteConstants";
+import { useRouter } from "vue-router";
 import RequestAssessment from "@/components/aest/students/assessment/Request.vue";
 import HistoryAssessment from "@/components/aest/students/assessment/History.vue";
 
@@ -32,9 +36,22 @@ export default {
       required: true,
     },
   },
-  setup() {
+  setup(props: any) {
+    const router = useRouter();
+    const goToStudentAppeal = (appealId: number) => {
+      router.push({
+        name: AESTRoutesConst.STUDENT_APPEAL_REQUESTS_APPROVAL,
+        params: {
+          studentId: props.studentId,
+          applicationId: props.applicationId,
+          appealId,
+        },
+      });
+    };
+
     return {
       AESTRoutesConst,
+      goToStudentAppeal,
     };
   },
 };
